src/requirements.txt
fastapi
uvicorn
gunicorn
Flask
python-dotenv
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
lxml
atlassian-python-api
requests
PyPDF2
faiss-cpu
azure-identity
azure-keyvault-secrets
azure-search-documents
azure-storage-blob
traceloop-sdk
flask-cors